LayoutTests/accessibility/placeholder-expected.txt

  
This test makes sure that the placeholder is returned as the correct attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS fieldElement.stringAttributeValue('AXPlaceholderValue') is 'search'
PASS pass.stringAttributeValue('AXPlaceholderValue') is 'Password'
PASS successfullyParsed is true

TEST COMPLETE
